requirements.txt
django
gunicorn
python-dotenv

# Rest API
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

djangorestframework_simplejwt
